test: move inline output assertions into options-3
