+++ packages/kit/test/apps/options-3/src/service-worker.js
