+++ packages/kit/test/apps/options-3/svelte.config.js
@@ -3,6 +3,9 @@ const config = {
 	kit: {
 		output: {
 			bundleStrategy: 'inline'
+		},
+		serviceWorker: {
+			register: false
 		}
 	}
 };
