+++ packages/kit/test/apps/options-3/vite.config.js
@@ -4,7 +4,8 @@ import { sveltekit } from '@sveltejs/kit/vite';
 /** @type {import('vite').UserConfig} */
 const config = {
 	build: {
-		minify: false
+		minify: false,
+		sourcemap: true
 	},
 	clearScreen: false,
 	plugins: [sveltekit()],
