--- packages/kit/test/prerendering/inline/src/routes/+layout.js
@@ -1,1 +1,0 @@
-export const prerender = true;
