--- packages/kit/test/prerendering/inline/src/routes/+page.svelte
@@ -1,7 +1,0 @@
-<h1>everything is inlined</h1>
-
-<style>
-	h1 {
-		color: teal;
-	}
-</style>
