--- packages/kit/test/prerendering/inline/tsconfig.json
@@ -1,8 +1,0 @@
-{
-	"compilerOptions": {
-		"allowJs": true,
-		"checkJs": true,
-		"noEmit": true
-	},
-	"extends": "./.svelte-kit/tsconfig.json"
-}
