--- packages/kit/test/prerendering/inline/vite.config.js
@@ -1,24 +1,0 @@
-import * as path from 'node:path';
-import { sveltekit } from '@sveltejs/kit/vite';
-
-/** @type {import('vite').UserConfig} */
-const config = {
-	build: {
-		minify: false,
-		sourcemap: true
-	},
-
-	clearScreen: false,
-
-	logLevel: 'silent',
-
-	plugins: [sveltekit()],
-
-	server: {
-		fs: {
-			allow: [path.resolve('../../../src')]
-		}
-	}
-};
-
-export default config;
